+++ Src/Processor/Taiko/TaikoScore.h
@@ -10,7 +10,7 @@ public:
 	TaikoScore(
 		s64 scoreId,
 		EGamemode mode,
-		s32 userId,
+		s64 userId,
 		s32 beatmapId,
 		s32 score,
 		s32 maxCombo,
